Add smoke test for MoveMap
