+++ tmol/tests/conftest.py
@@ -37,6 +37,14 @@ from .chemical import (  # noqa: F401
     rts_disulfide_res,
 )
 
+from .kinematics import (  # noqa: F401
+    ff_2ubq_6res_H,
+    ff_3_jagged_ubq_465res_H,
+    ff_3_jagged_ubq_465res_star,
+    ff_2ubq_6res_U,
+    ff_2ubq_6res_K,
+)
+
 from .torch import torch_device, torch_backward_coverage  # noqa: F401
 
 from .numba import numba_cudasim, numba_cuda_or_cudasim  # noqa: F401
